Moving some code around to make the Goldilocks crate closely match the other field crates.
